// src/modules/email/enums/email-type.enum.ts
export enum EmailTypeEnum {
  WELCOME = 'welcome',
  FORGOT_PASSWORD = 'forgot-password',
  LOGOUT = 'logout',
  OLD_VISIT = 'old-visit',
  NOTIFY_MISSING_PRODUCER = 'notify-missing-producer',
  NOTIFY_MISSING_MODEL = 'notify-missing-model',
  REPORT_PRODUCER_RECEIVED = 'report-producer-received',
  REPORT_MODEL_RECEIVED = 'report-model-received',
  NOTIFY_BANNED_ADVERTISEMENT = 'notify-banned-advertisement',
}